Fix comparison between signed and unsigned integers

+++ src/game/Entity.cpp
@@ -228,7 +228,7 @@ Entity::~Entity() {
 	ReleaseScript(&script);
 	ReleaseScript(&over_script);
 	
-	for(long n = 0; n < MAX_ANIMS; n++) {
+	for(size_t n = 0; n < MAX_ANIMS; n++) {
 		if(anims[n]) {
 			EERIE_ANIMMANAGER_ReleaseHandle(anims[n]);
 			anims[n] = NULL;
